Increased usability of built-in plugin configuration widgets.
Some of the plugin path fields still didn't adjust the style sheets when the user gives an invalid path. I think this is helpful for the user to see whether their path has been recognized or not. I have also added an extra call to validate() in imagesourcestep because the configuration fields weren't being validated on opening the widget.